+++ src/ios/APPLocalNotification.h
@@ -38,7 +38,7 @@
 - (void) getScheduledIds:(CDVInvokedUrlCommand*)command;
 // Informs if the app has the permission to show notifications
 - (void) hasPermission:(CDVInvokedUrlCommand*)command;
-// Ask for permission to show notifications
-- (void) promptForPermission:(CDVInvokedUrlCommand*)command;
+// Registers permission to show notifications
+- (void) registerPermission:(CDVInvokedUrlCommand*)command;
 
 @end